// src/modules/marketing/application/dtos/create-upcoming-launch.dto.ts
import type { Express } from 'express';

export class CreateUpcomingLaunchDto {
  category_label: string;
  title: string;
  start_date: string;
  image?: Express.Multer.File;
  image_url?: string;
  image_public_id?: string;
  link_url?: string;
  display_order?: number;
  status?: 'active' | 'inactive';
}